Increased memory sizes

+++ helpers.h
@@ -13,7 +13,7 @@
 // Function declarations
 void     debugPrintf(const char *, ...);
 uint32_t stringIndexFromSectionIndex(uint8_t *, uint8_t);
-void     addSectionData(SECTION *, uint8_t *, uint16_t);
+void     addSectionData(SECTION *, uint8_t *, uint32_t);
 int      get_matching_bracket(INSTRUCTIONS *, int);
 
 #endif // HELPERS_H_
